+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicNotificationService.java
@@ -6,11 +6,9 @@ import com.sprint.mission.discodeit.dto.data.UserDto;
 import com.sprint.mission.discodeit.entity.Notification;
 import com.sprint.mission.discodeit.entity.ReadStatus;
 import com.sprint.mission.discodeit.entity.User;
-import com.sprint.mission.discodeit.event.S3UploadFailedEvent;
 import com.sprint.mission.discodeit.event.MessageCreatedEvent;
 import com.sprint.mission.discodeit.event.RoleUpdatedEvent;
-import com.sprint.mission.discodeit.exception.DiscodeitException;
-import com.sprint.mission.discodeit.exception.ErrorCode;
+import com.sprint.mission.discodeit.event.S3UploadFailedEvent;
 import com.sprint.mission.discodeit.exception.channel.ChannelNotFoundException;
 import com.sprint.mission.discodeit.exception.notification.NotificationNotFoundException;
 import com.sprint.mission.discodeit.exception.user.UserNotFoundException;
@@ -19,19 +17,17 @@ import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.NotificationRepository;
 import com.sprint.mission.discodeit.repository.ReadStatusRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
-import com.sprint.mission.discodeit.security.DiscodeitUserDetails;
 import com.sprint.mission.discodeit.service.NotificationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.Cacheable;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 @Service
@@ -45,14 +41,14 @@ public class BasicNotificationService implements NotificationService {
     private final ReadStatusRepository readStatusRepository;
     private final ChannelRepository channelRepository;
     private final UserRepository userRepository;
+    //
+    private final CacheManager cacheManager;
 
     @Override
     @Transactional(readOnly = true)
-    @Cacheable("notifications")
-    public List<NotificationDto> getNotifications() {
-        UUID receiverId = getCurrentUserId();
+    @Cacheable(value = "notifications", key = "#receiverId")
+    public List<NotificationDto> getNotifications(UUID receiverId) {
         log.debug("알림 조회 시작: receiverId={}", receiverId);
-
         List<NotificationDto> notifications = notificationRepository.findByReceiverId(receiverId)
                 .stream()
                 .map(notificationMapper::toDto)
@@ -64,19 +60,20 @@ public class BasicNotificationService implements NotificationService {
 
     @Override
     @Transactional
-    @CacheEvict(cacheNames = "notifications", allEntries = true)
     public void deleteNotification(UUID notificationId) {
         log.debug("알림 삭제 시작: id={}", notificationId);
-        if (!notificationRepository.existsById(notificationId)) {
-            throw NotificationNotFoundException.withNotificationId(notificationId);
-        }
-        notificationRepository.deleteById(notificationId);
+        Notification notification = notificationRepository.findById(notificationId)
+                .orElseThrow(() -> NotificationNotFoundException.withNotificationId(notificationId));
+
+        UUID receiverId = notification.getReceiver().getId();
+        notificationRepository.delete(notification);
+
+        Objects.requireNonNull(cacheManager.getCache("notifications")).evict(receiverId);
         log.info("알림 삭제 완료 id={}", notificationId);
     }
 
     @Override
     @Transactional(propagation = Propagation.REQUIRES_NEW)
-    @CacheEvict(cacheNames = "notifications", allEntries = true)
     public void createMessageNotification(MessageCreatedEvent event) {
         log.debug("메시지 생성 알람 생성 시작 event={}", event);
         MessageDto messageDto = event.messageDto();
@@ -102,29 +99,30 @@ public class BasicNotificationService implements NotificationService {
             String title = authorUsername + " (#" + channelName + ")";
             String content = messageDto.content();
             Notification notification = new Notification(receiver, title, content);
-
-            log.debug("메시지 생성 알람 생성 성공: id={}", notification.getId());
             notificationRepository.save(notification);
+
+            Objects.requireNonNull(cacheManager.getCache("notifications")).evict(receiver.getId());
+            log.debug("메시지 생성 알람 생성 성공: id={}", notification.getId());
         }
     }
 
     @Override
     @Transactional(propagation = Propagation.REQUIRES_NEW)
-    @CacheEvict(cacheNames = "notifications", allEntries = true)
     public void createRoleUpdatedNotification(RoleUpdatedEvent event) {
         log.debug("권한 변경 알람 생성 시작: event={}", event);
         User receiver = event.user();
+
         String title = "권한이 변경되었습니다.";
         String content = event.oldRole().name() + " -> " + event.newRole().name();
         Notification notification = new Notification(receiver, title, content);
-        log.debug("권한 변경 알람 생성 성공: id={}", notification.getId());
-
         notificationRepository.save(notification);
+
+        Objects.requireNonNull(cacheManager.getCache("notifications")).evict(receiver.getId());
+        log.debug("권한 변경 알람 생성 성공: id={}", notification.getId());
     }
 
     @Override
     @Transactional(propagation = Propagation.REQUIRES_NEW)
-    @CacheEvict(cacheNames = "notifications", allEntries = true)
     public void createS3UploadFailedNotification(S3UploadFailedEvent event) {
         log.debug("바이너리 컨텐츠 생성 실패 알람 생성 시작: event={}", event);
         User receiver = userRepository.findByUsername("admin")
@@ -142,26 +140,9 @@ public class BasicNotificationService implements NotificationService {
                         "Error: " + errorMessage;
 
         Notification notification = new Notification(receiver, title, content);
-        log.debug("바이너리 컨텐츠 생성 실패 알람 생성 성공: id={}", notification.getId());
-
         notificationRepository.save(notification);
-    }
 
-    private UUID getCurrentUserId() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        if (authentication == null || !authentication.isAuthenticated()) {
-            throw new DiscodeitException(ErrorCode.INVALID_TOKEN);
-        }
-
-        Object principal = authentication.getPrincipal();
-
-        if (!(principal instanceof DiscodeitUserDetails userDetails)) {
-            throw new DiscodeitException(ErrorCode.INVALID_USER_DETAILS);
-        }
-
-        UserDto userDto = userDetails.getUserDto();
-        log.debug("권한 변경 알람 생성 완료: id={}, newRole={}", userDto.id(), userDto.role().name());
-        return userDto.id();
+        Objects.requireNonNull(cacheManager.getCache("notifications")).evict(receiver.getId());
+        log.debug("바이너리 컨텐츠 생성 실패 알람 생성 성공: id={}", notification.getId());
     }
 }
